+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `PieceList` copy by reference.
 
+### Changed
+
+- **BREAKING** `Piece.max` now `.max_board` for clarity.
+
 ## [0.1.2] - 2024-08-11
 
 ### Added